+++ environments/ipex.yml
@@ -25,10 +25,8 @@ dependencies:
   - ffmpeg
   - pip:
     - --extra-index-url https://pytorch-extension.intel.com/release-whl/stable/xpu/us/
-    - torch==2.0.1a0; sys_platform == 'linux'
-    - torch==2.0.0a0; sys_platform == 'win32'
-    - intel_extension_for_pytorch==2.0.110+xpu; sys_platform == 'linux'
-    - intel_extension_for_pytorch==2.0.110+gitba7f6c1; sys_platform == 'win32'
+    - torch==2.1.0a0
+    - intel-extension-for-pytorch==2.1.10+xpu
     - openvino
     - onnxruntime-openvino
     - flask-cloudflared==0.0.10
